게임매니저 씬 붙이기 완료, 타임보너스 오버플로우 남

+++ New Unity Project/Assets/Scripts/GameManager.cs
@@ -4,12 +4,12 @@ using UnityEngine.SceneManagement;
 public class GameManager : MonoBehaviour
 {
     public float scrollSpeed { get; set; } = 0.5f;
-    public int score { get; set; } = 0;
+    public long score { get; set; } = 0;
     public int life { get; set; } = 4;
     void Awake() {
         //SceneManager.LoadScene("MainMenu", LoadSceneMode.Additive);
     }
-    public void AddScore(int score) {
+    public void AddScore(long score) {
         this.score += score;
     }
 }
